DIANA-Personal Assistant Using Speech Recognition
Diana is a personal assistant from the game Hitman which inspired me to further develop my own personal assistant, it uses the speech engine that can be downloaded or is in built in some visual studio and recognize the speech spoken by us (commands) and convert it to machine language and further do things in the foreground whereas in the background coding’s and conditions are the main backbone of each process.
Many companies around the globe like Apple, Microsoft and Nuance are developing their own preferred personal assistant like Siri, Google talk and Cortana to name a few.
Speech Recognition-Speech recognition is the ability of a machine or program to identify words and phrases in spoken language and convert them to a machine-readable format.
As we see that speech is a very important part of the new world programming. (Nuance) a company have come up with voice biometrics where only a person’s voice is needed for password and login procedures. It is a new step for programmers around the globe to combine voice, speech with different programming languages and develop software like Siri and Cortana.